pyproject.toml
[tool.poetry]
name = "trip-planner-ai"
version = "0.1.0"
description = "Simple ai trip planner"
authors = ["Watcharaphon Piamphuetna"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
langchain-openai = "^0.0.5"
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"
flask = "^3.0.2"
flask-cors = "^4.0.0"
langchain-google-genai = "^0.0.11"
crewai = "^0.22.5"
markdown = "^3.6"
langchain-anthropic = "^0.1.4"
langchain-nvidia-ai-endpoints = "^0.0.4"
langchain = "^0.1.13"
langchain-community = "^0.0.29"
duckduckgo-search = "^5.2.1"
litellm = {extras = ["proxy"], version = "^1.34.18"}

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"